+++ CHANGELOG.md
@@ -4,9 +4,9 @@
 
 - Editor
   - Add support for LSP feature `textDocument/selectionRange`. #1961
-
 - API/CLI
   - Replace `tools.cli` with `babashka.cli`. #2036
+  - `diagnostics`: Add --diff option to limit diagnostics to a git revision range. #2029
 
 ## 2025.04.23-18.16.46